[Frontend] Support for chat completions input in the tokenize endpoint (#5923)
